Rename snapshot_keyper package
